split player to different files
